static_cast<int> possible.size()
On Win64, it's a long long, while it's merely a long on Linux amd64.
So, static_cast it to "int", so I can use %d instead of switching
%ld or %lld depending on whether it's LLP64 or LP64.

+++ landlock.cpp
@@ -410,7 +410,7 @@ EX eLand pickLandArc() {
       printf("Adding %s\n", linf[i].name);
     }
   }
-  printf("%ld lands to pick from!\n", possible.size());
+  printf("%d lands to pick from!\n", static_cast<int>(possible.size()));
   if (possible.size())
     return possible[hrand(possible.size())];
